// src/app/models/IAppUser.ts
export interface IAppUser {
    // id?: number | null;
    email?: string | null;
    username?: string | null;
    password?: string | null;
    first_name?: string | null;
    last_name?: string | null;
    // firstName?: string | null;
    // about?: string | null;
    // isStaff?: boolean | null;
    // isActive?: boolean | null;
    // token?: string | null;
    // accesToken?: string | null;
}